+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ This document contains the FSL-MRS release history in reverse chronological orde
 2.1.7 (WIP)
 ---------------------------
 - Added `fsl_mrs_proc mrsi-align` which can perform frequency and phase alignment across voxels using cross correlation.
+- Added `fsl_mrs_proc mrsi-lipid` which can perform lipid removal using Bilgic et al's L2-regularised method.
 - `fsl_mrs_proc fshift` can now take NIfTI images of matched shape witht he `--shifthz` anf `--shiftppm` to apply per-voxel shifts.
 - Fixed bug with incorrectly calculated zero order phase when given in degrees.
 - Improved interface and options for dynamic fitting driven preprocessing.